Move ValueObjectAssertions to src, use from provided testing specifications
